Pulled FileConverter and FileLoader classes out of the converter module
